+++ packages/api/package.json
@@ -59,6 +59,7 @@
     "@cumulus/db": "9.0.1",
     "@cumulus/earthdata-login-client": "9.0.1",
     "@cumulus/errors": "9.0.1",
+    "@cumulus/es-client": "9.0.1",
     "@cumulus/ingest": "9.0.1",
     "@cumulus/launchpad-auth": "9.0.1",
     "@cumulus/logger": "9.0.1",
@@ -69,6 +70,7 @@
     "@mapbox/dyno": "^1.4.2",
     "aggregate-error": "^3.1.0",
     "ajv": "^6.12.3",
+    "aws-sdk": "^2.585.0",
     "aws-serverless-express": "^3.3.5",
     "body-parser": "^1.18.3",
     "commander": "^2.15.0",
