+++ packages/es-client/.gitignore
@@ -4,7 +4,6 @@ dist/
 test-reports
 tmp
 .tmp
-/*.js
 !/nyc.config.js
 /*.d.ts
 /lib/*.js
