+++ packages/es-client/package.json
@@ -9,9 +9,6 @@
   "engines": {
     "node": ">=12.18.0"
   },
-  "files": [
-    "*.js"
-  ],
   "publishConfig": {
     "access": "public"
   },
@@ -41,6 +38,7 @@
     "@cumulus/message": "9.0.1",
     "@elastic/elasticsearch": "^5.6.20",
     "lodash": "~4.17.20",
-    "moment": "2.24.0"
+    "moment": "2.24.0",
+    "p-limit": "^1.2.0"
   }
 }
