+++ packages/es-client/testUtils.js
@@ -1,14 +1,13 @@
 const { randomString } = require('@cumulus/common/test-utils');
 
-const bootstrap = require('../lambdas/bootstrap');
-
+const { bootstrapElasticSearch } = require('./bootstrap');
 const { Search } = require('./search');
 
 const createTestIndex = async () => {
   const esIndex = randomString();
   const esAlias = randomString();
   process.env.ES_INDEX = esIndex;
-  await bootstrap.bootstrapElasticSearch('fakehost', esIndex, esAlias);
+  await bootstrapElasticSearch('fakehost', esIndex, esAlias);
   const esClient = await Search.es('fakehost');
   return { esIndex, esClient };
 };
